requirements.txt
# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Redis caching
redis==5.0.1

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# AI/ML
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0

# Utilities
python-dotenv==1.0.0
